Modified, nearly, every class to move away from using an ArrayList for holding particles to using a fixed-size Array for holding the particles.
The benefits of this change are that there will never be an absurd amount of particles being created and thus the CPU/RAM usage should, in theory, be reduced. All of the effects have been, roughly, tuned to have a specific number of particles allowed on the screen at any given time.

One other change that was made is that the Effects now only update half of the particles on screen during each update call. The updated particles alternate between the first half of the Array of particles and the second half of the Array of particles with every third call being discarded. This, simple, change has shown a drastic decrease in CPU usage.