Modified, nearly, all classes.
The particle system no-longer creates a new object for every particle and destroys old objects for dead particles.
The system creates an object for each of the particles up to the total number of particles in the effect.

Whenever a particle "dies" it's object is reused for a new particle.